+++ CHANGELOG.md
@@ -458,6 +458,8 @@
     - Added HealingSource, an interface applied to all sources of healing from player actions, for easier tracking of healing sources.
   - Move Change
   - Shiny Chance Calculation Event (with player context)
+  - Dex Information Changed; this event is fired when the Pokédex is updated with new information.
+    - Added DexDataSource, a marker interface for data sources that can be used to update the Pokédex - currently just used in Pokemon and PokedexEntityData
 - Rebuilt large swaths of the model animation code to simplify it.
 - Renamed a bunch of things from %Poseable% to %Posable% because spelling.
 - Renamed StatelessAnimation to PoseAnimation.